requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3
numpy==1.24.3